requirements.txt
nvidia-ml-py
pillow